apps/backend/consent-ingestion/requirements.txt
fastapi
uvicorn
sqlalchemy
pg8000
python-dotenv
pytesseract
Pillow
pyzbar
pypdf
langdetect
python-multipart
